Move get_obj_from_str() and instantiate_from_config() into package
